DDCE-2562 - Remove duplicate code
